Ch. 17§05 initial edits on 'Composing Streams'
