--- listings/ch17-async-await/listing-17-45/src/main.rs
@@ -1,35 +1,0 @@
-use std::{pin::pin, time::Duration};
-
-use trpl::{ReceiverStream, Stream, StreamExt};
-
-fn main() {
-    trpl::block_on(async {
-        let mut messages =
-            pin!(get_messages().timeout(Duration::from_millis(200)));
-
-        while let Some(result) = messages.next().await {
-            match result {
-                Ok(message) => println!("{message}"),
-                Err(reason) => eprintln!("Problem: {reason:?}"),
-            }
-        }
-    })
-}
-
-fn get_messages() -> impl Stream<Item = String> {
-    let (tx, rx) = trpl::channel();
-
-    // ANCHOR: messages
-    trpl::spawn_task(async move {
-        let messages = ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"];
-        for (index, message) in messages.into_iter().enumerate() {
-            let time_to_sleep = if index % 2 == 0 { 100 } else { 300 };
-            trpl::sleep(Duration::from_millis(time_to_sleep)).await;
-
-            tx.send(format!("Message: '{message}'")).unwrap();
-        }
-    });
-    // ANCHOR_END: messages
-
-    ReceiverStream::new(rx)
-}
